シーンファイル読み書きの処理を書いた。動作確認はまだ。シーンの名前変更と、シーンの作成、ロードの仕組みを変更。

+++ TenkikunEngineProject/FilePrintRect.cpp
@@ -147,7 +147,7 @@ void FilePrintRect::LoadFoler()
 		}
 	}
 
-	int row = fileIcons.size() / maxFileNumInRow;
+	int row = (int)fileIcons.size() / maxFileNumInRow;
 	if ((int)fileIcons.size() % maxFileNumInRow != 0) {
 		row++;
 	}
